+++ gensim/models/experimental/__init__.py
@@ -1,0 +1,3 @@
+from .dssm import DSSM  # noqa:F401
+from .drmm_tks import DRMM_TKS  # noqa:F401
+from .drmm import DRMM  # noqa:F401
