+++ gensim/similarity_learning/__init__.py
@@ -1,5 +1,3 @@
-from .models import DSSM, DRMM_TKS  # noqa:F401
-from .preprocessing import WikiQAExtractor, QuoraQPExtractor, WikiQA_DRMM_TKS_Extractor, ListGenerator  # noqa:F401
 from .custom_losses import rank_hinge_loss  # noqa:F401
 from .custom_callbacks import ValidationCallback  # noqa:F401
 from .evaluation_metrics import mapk, mean_ndcg  # noqa:F401
