--- gensim/similarity_learning/models/__init__.py
@@ -1,2 +1,0 @@
-from .dssm import DSSM  # noqa:F401
-from .drmm_tks import DRMM_TKS  # noqa:F401
